Store tip amounts as whole cents
SQLite has no decimal type, so a decimal was kept as TEXT: exact to read back
but lexicographic, which meant '100.00' sorted before '9.00' and the SUM
behind a singer's tip total had to coerce text through floating point to get
a number. Cents in an INTEGER sort and sum correctly and never round.

Tip.Amount becomes Tip.AmountInCents, and the total returned by the
repository and service is cents too. Presentation goes through
KHost.Domain.Monetary.MoneyExtensions.ToCurrency(), so 1250 renders as
$12.50 in the tips table, the singer panel and the delete confirmation. The
amount input already worked in cents at the JS boundary, so the model now
matches it and the decimal round-trip in between is gone.

The migration is hand-written: scaffolding it as drop-then-add would have
discarded every recorded amount. It carries values across with ROUND before
the integer cast, because multiplying by 100 in floating point leaves 0.29 as
28.999... and a cast truncates. TipAmountConversionTests proves that against
a database migrated to the previous version and seeded with the values that
go wrong - 0.29, 0.01, 3.33, 99999.99 - and Down round-trips them back.

+++ src/KHost.Abstractions/Repositories/ITipsRepository.cs
@@ -5,5 +5,6 @@ using KHost.Abstractions.Models;
 public interface ITipsRepository : IRepository<Tip>
 {
     Task<IReadOnlyList<Tip>> GetByUserIdAsync(Guid userId);
-    Task<decimal> GetTotalByUserIdAsync(Guid userId, DateTime? from = null, DateTime? to = null);
+    /// <summary>Total in whole cents.</summary>
+    Task<int> GetTotalByUserIdAsync(Guid userId, DateTime? from = null, DateTime? to = null);
 }
